refactor: cohort 관련 익셉션 계층 이동
