+++ src/main/kotlin/com/server/dpmcore/cohort/application/exception/CohortExceptionCode.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.cohort.domain.exception
+package com.server.dpmcore.cohort.application.exception
 
 import com.server.dpmcore.common.exception.ExceptionCode
 import org.springframework.http.HttpStatus
